refactor: upd folders structure
